Sprint 2 - Basic mass spring started
Add Mass-Spring game object, this adds the fragments as EDEMElements to the scene and connects them to their neighbours through pore springs.

Add Edem Element simple, object for a fragment has a mass,  and reference to any pore springs attached. Calculates the acceleration from the pore spring forces and adds the acceleration to the rigidbody!

Add Pore spring class , has reference to its two edem elements it is attached to.  When called by the mass spring system it calculates the force being loaded on each element attached.

Minor changes to camera control script