Sprint 3 - FEM - Prep
Small Refactor - Moved Scripts into the folders for their respective models, so it is easier to find each

FireCannon Script - Corrected issue where mesh didn't match the size of the collider

FEM - Model performs calculations by storing data in matrices, so basic Matrix class was added.

Matrix class - Added operator overloads for addition, subtraction, multiplication, as well as for multiplication with a custom Vector class

Matrix class - A determinant method that works for nxn matrices, an inverse method that works for nxn matrices

Matrix class - QR decomposition method, that is used in calculating the eigenvalues and eigenvectors, so that the polar decomposition can be calculated which is necessary for the calculation of the FEM method(removes rotation from the deformation gradient, for calculating the linear strain) Polar Decomp still to be done!

A basic vector class for use in the QR and eigen calculations, basic indexer, constructors, dot product method, magnitude and normalised methods. Can be instantiated by passing a Matrix and the column of that matrix.
It is very bare bones, just covering what i need for these calculations.