Sprint 3 FEM - Stress calculations
Added calculations for strain, using strain to find stress and finally calculating the force acting on each node.

Next step is calculating Jacobians of the force, these can be used to get the stiffness matrix.

The force can be used for calculating the separation tensor to find if fracture has occurred