Sprint 3 - FEM - Minor Changes
Added poissons Ratio, youngs modulus for calculation of material paramaters lambda and mu.

Altered createInitialMesh so that it creates the FEM elements and adds them to the list.

Removed vertices, surface triangles and elements object wide variables, only used in createInitialMesh where they are needed.

Added list of FEMElements to FEM script, may need to add and remove elements could possibly be changed to an array